Fix express parameter parsing
Express doesn't give us a way to escape ':' in a route. Google's api library uses ':'
in the requested URL. So we have to choose between invalid parsed params or breaking
the existing google play client by returning a 404. That is not ok:

/com.example.blah/:editId:whatever

As a HACKY workaround, we know editIds are always just digits, so pass
a custom regex to that param defintion that stops at any other character.

The final ":whatever" at the end of the URL still matches and a lame parameter
is parsed, but we can just ignore it.

This works because so far all the /com.example.blah/99999999:whatever URLs
always have the editId adjacent `:whatever`